Book_Store_BE/book-service/requirements.txt
django
djangorestframework
celery
redis
